wm: Make WM_MOVE_PANEL use top-right coordinates.
It was poorly documented in the header, but this message
type previously expected Chrome to pass the window manager
the displaced coordinates of the top-left corner of the
panel's titlebar window.  This doesn't work if it's possible
for the panel's titlebar window to get resized during the
drag, since the panel's titlebar is right-justified with the
content window.  I plan to make us resize the titlebar mid-
drag soon to support docking panels on the sides of the
screen.

This change makes the window manager expect the message to
contain the displaced coordinates of the top-right corner of
the titlebar window instead.  (It's technically one pixel to
the right of the corner -- see wm_ipc.h for details.)  It'll
require a corresponding change in Chrome.

It also adds (crude) versioning for WM/Chrome IP messages,
so this change should still work with existing Chrome builds.

Review URL: http://codereview.chromium.org/527005