wm: Reduce jitter in active window property.
When unmapping a transient window, we were previously setting
the property to None before setting it to the newly-focused
toplevel window, which could make the toplevel window's frame
flash inactive/active.

This change makes us just set the property to the toplevel
window even before we've received the FocusIn event.  We
should probably do the same thing in other places -- trying
to accurately decipher the meaning of all of the FocusIn and
FocusOut events that we receive seems hopeless.

BUG=chromium-os:1556
TEST=included

Review URL: http://codereview.chromium.org/600092