wm: Let panels in the panel bar be dragged up and down.
If a collapsed panel is dragged more than halfway up, we expand
it, and vice versa for already-expanded panels.

I use the initial direction of the drag to determine if the
panel is being moved horizontally or vertically.  It feels a
little restrictive; perhaps the drag axis should be determined
based on the last few motion events.

BUG=none
TEST=included

Review URL: http://codereview.chromium.org/661057