wm: Try to avoid some crashes when windows disappear.
Tidy looks like it has a few races where it's trying to do
something with windows after they've already gone away.
With the exception of one CHECK() that shouldn't be there,
these would be harmless, except that they can produce errors
that Xlib ends up seeing in the event queue, leading to
(extremely frequent) crashes.

One fix would be moving away from Xlib entirely.  This
probably isn't going to happen anytime soon, since there's a
non-trivial amount of stuff that hasn't been ported to XCB
or seems to be broken, at least in the version of XCB that
we're using.

The way that we should probably work around this (when we're
not trying to hit an internal milestone) is to install a
permanent Xlib error handler, getting rid of the existing
code that we use to bracket Xlib calls.  This will diminish
our error-reporting to some degree (since we won't know
which request generated the error), but should avoid this
source of crashes.  There may be some happy medium where we
have a generic error handler that we use most of the time
but swap in a different one to bracket Xlib calls.

TEST=clicked on windows for a few minutes
BUG=chromium-os:1624, maybe

Review URL: http://codereview.chromium.org/660237