wm: Add an 'xid' field to WmIpc::Message.
This contains the 'window' field from the original ClientMessage
event, and should free up an additional parameter in the message
payload for new messages (i.e. no need to use one of the parameters
to refer to the window that is receiving the event now).

Note that this change still needs to be added to Chrome before we
can use this in messages sent from the WM to Chrome.

BUG=none
TEST=included

Review URL: http://codereview.chromium.org/661209