wm: Support fullscreen mode for panels.
This is a bit tricky.  While a panel is fullscreen,
changes can be made (like other panels being closed)
than affect where the panel should end up when its
original state is restored.  While it's fullscreen,
I record changes made to it so they can be applied
later.  The alternative approaches that I could think
of (like keeping a fake window around that can be
swapped in in place of the real content window in the
Panel object while it's fullscreen) seemed even
hackier.

This is still a bit rough (for instance, we should
probably unfullscreen the panel if the user hits
F12 to switch to overview mode, but that'll require
making LayoutManager notify WindowManager so it can
notify PanelManager), but it's hopefully a good
start for now.

BUG=chromium-os:1850
TEST=included

Review URL: http://codereview.chromium.org/890006