wm: Add support for groups of key bindings.
I'm doing this as a precursor to a change to make the
WM listen for the volume up and down keys -- we have
some bindings that should always be present and other
ones that should only be present when the user is
logged in.  This also fixes a long-standing TODO to
clean up some repetitive code that would occasionally
get out-of-sync for adding and removing bindings for
the different layout manager modes.

BUG=chromium-os:1880
TEST=included

Review URL: http://codereview.chromium.org/1003003