wm: Handle missing keycodes more gracefully.
We used to DCHECK if a combo that we had previously
registered used a keysym that was missing its keycode after
a keymap change, or avoid registering a combo if it was
initially missing a keycode (even if a later keymap change
made it usable).  Lots of weird things appear to be
happening keymapping-wise right after the X server starts,
and this is causing problems with the login controller code.
Hopefully this change will fix them.

BUG=none
TEST=included

Review URL: http://codereview.chromium.org/1148001