wm: Write file when first Chrome window is mapped.
If a UI test needs to log in first, it's currently
tough for it to tell when stuff has settled and it's
safe to start testing without fear of being
interrupted by the initial Chrome window sliding
onscreen.

Off by default; I'll check in an init script change
next (do we just want this for test images, or could
it be useful elsewhere?).

BUG=chromium-os:2264
TEST=tried it

Review URL: http://codereview.chromium.org/1366003