wm: Grab the server to avoid some races.
The races were previously worked around, but this is
a cleaner approach.  We couldn't do this before because
grabs could cause deadlocks with Clutter, which was
using its own X connection.  Now that we're using our
own implementation that shares a connection with us,
this is no longer a concern.

BUG=none
TEST=tried it

Review URL: http://codereview.chromium.org/1574006