wm: Update active window property when panels are focused.
When a new panel was created, it looks like we were
misinterpreting a FocusIn event that made us think that the
toplevel window was getting focused.  This change makes us
honor a later FocusIn event for the panel, so that the
active window property ends up pointing at it.  A better fix
that I'll probably do soon is to ignore FocusIn and FocusOut
events and instead just track the focus based on where we
assigned it -- focus events are pretty much useless. :-(

BUG=chromium-os:2368
TEST=tried it

Review URL: http://codereview.chromium.org/1584012