wm: Don't resize actors until underlying window is resized.
Rather than resizing a window's TexturePixmapActor at the
same time as when we ask the X server to resize the client
window, we defer it until we get a ConfigureNotify event
saying that the window has been resized.  This looks like
it's sufficient to get rid of the weird scaling effect that
would happen otherwise, where the window's previous
contents are stretched or compressed to fit in the new size.

BUG=chromium-os:1279
TEST=included. works on my X200; still need to test on Eee

Review URL: http://codereview.chromium.org/1578010