wm: Rewrite focus code.
The old code was based on FocusIn and FocusOut events,
which leads to race-prone, confusing code.  Since it's
unlikely that any clients other than the window manager
will be assigning the focus, we just keep track of the
last window that we focused.  I also added a new
FocusManager class that handles things like updating the
_NET_ACTIVE_WINDOW hint and adding and removing button
grabs on windows so they don't need to be reimplemented
in LayoutManager, LoginController, etc.

BUG=none
TEST=updated existing tests and wrote some new ones

Review URL: http://codereview.chromium.org/2020008