wm: Move transient window code into its own class.
This pulls the TransientWindow struct and a bunch of the
supporting code out of ToplevelWindow and puts it all into a
TransientWindowCollection class.  I'm going to use the new
class to add transient window support to the Panel class.

This CL isn't intended to make any functional changes.

BUG=chromium-os:2455
TEST=existing tests pass

Review URL: http://codereview.chromium.org/2132020