wm: Fix a race condition by anticipating mapping.
It's possible for Chrome to do something entirely reasonable
(map/show a window and immediately send a message to us
about it) that results in us receiving a message about a
window before we're completely set up to handle it.

This change makes us set things up earlier, as soon as we've
forwarded the map request to the X server, to avoid the race
in most cases.  (It'd still be present for override-redirect
windows, but Chrome doesn't send us messages about those and
I can't think of any way to avoid the race there anyway.)

BUG=chromium-os:3444
TEST=included

Review URL: http://codereview.chromium.org/2118012