wm: Track logged-in state so we don't need to restart.
Make the window manager track the _CHROME_LOGGED_IN property
on the root window and do various things when it changes
(update the path to which we write logs, update key bindings,
hide login windows if they're still around after we see the
initial Chrome window, etc.).

Also get rid of the windowmanager_not_logged_in job, and
make the regular windowmanager job get started after X.
Pass some additional flags to it so it knows how to handle
both logged-in and not-logged-in states.  Get rid of "wm_"
prefix on flags declared in window_manager.cc -- tedbo
recommended doing this a long time ago, but it's not being
used consistently and just makes the names longer.  grep is
easy to use anyway.

BUG=chromium-os:2790,chromium-os:3187
TEST=included

Review URL: http://codereview.chromium.org/2117027