wm: Hide other windows when screen locker window is mapped.
This adds "visibility groups" to the compositor, a fairly
ad-hoc way to quickly hide or show a bunch of actors.  It
would maybe be more correct for the window manager to handle
hiding or showing the other windows itself, but trying to
track the states of all other windows while the screen locker
window is up and then restore them afterwards would be a
nightmare.  This approach seems like it may be useful for
other things in the future.

BUG=chromium-os:2852,chromium-os:2914
TEST=added

Review URL: http://codereview.chromium.org/2494004