initial add of chromeos shell and cros-term
BUG=chromium-os:3884
TEST=None yet.

cros-term takes care of launching the appropriate xterm stand-in, with a suitable environment.

crosh is a read-eval-print loop with a whitelisted set of commands.  It is intended to be started by our xterm, instead of bash, so that we don't have to give a full bash prompt.

I don't believe platform/window-manager is really the correct place to stash these scripts in the long term.  Short term, we'd like to land this in time to cherry-pick into R7, and window-manager seems like the best of the wrong places.

If this ends up being something we keep around, it probably deserves its own repo.

Review URL: http://codereview.chromium.org/2733001