wm: Remove old, unused metrics-reporting code.
The protobuf part of this code is giving me trouble
while I'm trying to update the window manager ebuild.
It sounds like this is all unused now (Chrome doesn't
do anything with the property that the window manager
sets), so it's probably simplest to delete it for now.
I'll remove the corresponding code from Chrome in a
separate change.

BUG=none
TEST=built it

Review URL: http://codereview.chromium.org/2762010