Changes to support portait overview mode and match UI mock.
This changes the layout manager so that it renders snapshots at 1.0 scale when they are face-on to the screen.  This is so that the snapshots are readable when displayed in overview mode.

Also included: changes to constants that Cole wanted, and we now stop when we hit the end of the snapshots instead of wrapping around to the beginning.   Also, animations of windows other than the selected snapshot window were removed.

TEST=built and ran window manager with modified Chrome on workstation and netbook.
BUG=none

Review URL: http://codereview.chromium.org/2757002