wm: Add support for (non-panel) fullscreen windows.
I used to basically ignore requests to fullscreen
normal windows, with the reasoning that they're
already covering the whole screen.  This change
improves support by restacking the window on top
of other windows (e.g. panels) and unfullscreening
it when it loses the focus.

BUG=chromium-os:529
TEST=added

Review URL: http://codereview.chromium.org/2500003