wm: Improve handling of panel resize requests.
This moves the responsibility for applying the requests from
the Panel class (which didn't really know if a request was
appropriate) to its PanelContainer, which fixes some
problems with docked panels (we now disallow width changes
for them, use the correct resize gravity, and repack the
other panels after a height change).

I'm also changing the minimum panel content dimensions to
200x200, at least for the time being.  Chrome (or more
likely, GTK+) appears to try to enforce this limit, so I
figure that it's better not to fight it.

BUG=chromium-os:4180
TEST=added

Review URL: http://codereview.chromium.org/2814023