wm: Clean up TexturePixmapActor implementation.
You can blame piman for this -- I was following his advice
to try to use pixmaps for getting the root window's contents
into the window manager instead of copying them to an image,
but I got confused by the current state of the
TexturePixmapActor code and had trouble modifying it to use
non-window pixmaps because of the way that a bunch of
window-specific code was implemented inside of the visitor
instead of in the compositor.  Here's my attempt at cleaning
things up.

This change:
- changes the TexturePixmapActor interface to make it more
  clear when the window manager needs to let it know that
  things have changed
- moves some work out of OpenGlPixmapData and into
  Window, like looking up X pixmaps and creating
  and modifying Damage objects
- simplifies the way that OpenGlPixmapData objects are used,
  so that TexturePixmapActor throws the current data away
  when a new pixmap is created, and then the visitor creates
  a new object when it sees that the old one is missing (I
  think that this is essentially how things worked before,
  but all of that work was done by a static method of
  OpenGlPixmapData instead)
- gets rid of some places where OpenGlPixmapData was
  modifying its actor

There's more work that can be done in this direction.  I
figured I'd wait for some feedback about whether this is an
improvement before touching the GLES stuff, so please let me
know what you think.

Review URL: http://codereview.chromium.org/2863010