wm: Use initial contents of root win as background for login.
We create a pixmap and copy the initial contents of the root
window into it.  This gets rid of a flash between the point
at which X has started (with the previous contents of the
framebuffer in the root window) and the point at which Chrome
has mapped its login background window.

BUG=chromium-os:3224
TEST=tried it

Review URL: http://codereview.chromium.org/2804013