wm: Add explicit GLES shaders.h dependency to build.
SCons doesn't figure this out on its own due to the format
of the include paths that we're using.  In the updated
ebuild files that I'm working on, this leads to a race where
SCons builds opengles_visitor.o in parallel with the Python
code that generates the shaders.h file that the visitor
includes.

BUG=none
TEST=built it
TBR=kwaters

Review URL: http://codereview.chromium.org/2817028