wm: Pass keycodes (vs. keysyms) to key bindings code.
I was originally doing this as part of a change to work
around an X server autorepeat bug, but it's since been
fixed.  This cleanup still seems worthwhile, though.  I'm
also leaving in a method that I added to permit peeking
ahead in the X event queue.

BUG=none
TEST=updated

Review URL: http://codereview.chromium.org/2623003