wm: Use slow path when texture-from-pixmap is unavailable.
When GLX_EXT_texture_from_pixmap is unavailable (typically
because we're using Mesa's software rasterizer instead of
hardware acceleration), we fall back to getting windows'
contents through GetImage requests to the X server. This is
quite slow but ought to let us run under VMs and in nested X
servers such as Xephyr, which is useful for developers.

This change also makes us display an annoying "unaccelerated
graphics" image in the upper-left corner to make it clear
what's going on.  (I guess that the graphics could still be
accelerated in this case, but they probably aren't if TFP
is missing.)

We can avoid an additional copy and make this faster by using
the MIT-SHM extension instead of GetImage, but I doubt that
it's worth the trouble at the moment.

BUG=chromium-os:3434
TEST=ran it under Xephyr

Review URL: http://codereview.chromium.org/2934008