wm: Support "independently positioned" panels.
(Since that is a pain to type, I'm calling them "floating
panels" in the source.)

When the user drags a panel to the left, this change makes
it just stick there now.  When the main block of "packed"
panels grows large enough to encroach on floating panels,
the floating panels get pushed to the left.  When the block
gets smaller, the floating panels return to their previous
locations.  When one floating panel gets dropped on top of
other floating panels, I try to do something reasonable to
get the other panels out of the way.

There are some things that still need work.  The "one
floating panel gets dropped on top of another floating
panel" stuff is sort of weird, in particular, but it's not
clear to me what the right behavior is.  This at least lets
users move panels so they don't block content and gives us
something to improve on, and doesn't seem to crash.  I
think.

BUG=chromium-os:3418
TEST=added

Review URL: http://codereview.chromium.org/3072002