wm: Support the _NET_WM_SYNC_REQUEST protocol.
This is described in more detail (but just barely) in
http://standards.freedesktop.org/wm-spec/wm-spec-latest.html,
but the general idea is that a client creates a counter for
its window.  When the window manager wants to resize the
window, it sends the client a message telling it a new
counter value before telling the X server to do the resize.
After the client has finished repainting the window in
response to the resize, it updates the counter.  After the
window manager is notified that the counter has been
updated, it composites the window's new contents.  This
should let us avoid compositing the resized window before
it's been repainted.

Note that in the case of non-override-redirect windows, this
can also be used to avoid compositing a newly-mapped window
before the client has had a chance to paint it: if we resize
a newly-created window before mapping it, then we can wait
for the sync notification from the client before fetching
its pixmap.  I've omitted this from Window::HandleMapNotify()
since it has a side effect: we now start animating windows
before we're actually drawing them, so e.g. the background
is briefly visible when a new window is sliding onscreen
before we get the window's contents and start drawing it.
I'll work on this in a following change.

BUG=chromium-os:5599
TEST=added

Review URL: http://codereview.chromium.org/1528015