wm: Improve login transitions.
This fixes a bug where the login controller would briefly
select the guest entry immediately before the initial
browser window was mapped, as a result of the other login
entries all getting unmapped.

It also makes us destroy the login controller after the
initial browser window gets created, and hold off on creating
the layout and panel managers until the user has logged in.
The latter part of this simplifies those classes a bit, by
making them not need to check whether the user is logged in
or not.

I also tried to clean up the logic about which event
consumers manage which windows.  Since there's a point in
time where both the login controller and the layout and panel
managers exist (after the _CHROME_LOGGED_IN property is set
but before the initial browser window has been mapped), the
only non-login-specific windows that the login controller
manages now are ones that are transient for login windows.  I
didn't notice any windows that are supposed to be shown
during login that are omitted by this.

Finally, it makes the window manager ignore WM_HIDE_LOGIN
messages.  I'll make Chrome stop sending these soon.

The initial browser window is still drawn as a blank
white window when it first comes up, but that's unrelated.
I've opened http://crosbug.com/5599 to track coming up with
a fix.

BUG=chromium-os:3224
TEST=tried it and added tests

Review URL: http://codereview.chromium.org/3126006