wm: Make panel titlebar cast shadow on content.
This makes us draw a faint shadow across the top of panel
content windows.  It is perhaps overkill to use the Shadow
class for this (the shadow only has one image, stretched
across the bottom), but it handles things like checking the
dimension of the image, cloning the texture instead of
loading it from disk each time, etc.

BUG=chromium-os:5165
TEST=added

Review URL: http://codereview.chromium.org/3183007