wm: Fix override-redirect shadow stacking bug.
We were adding shadow actors to the stage after their
corresponding windows, resulting in them being stacked above
the windows.  If the window got restacked, we'd put the
shadow in the correct place (underneath the window), but the
problem would persist for windows that never got
reconfigured.  I'm not sure when this was introduced; my
recent cleanup of the shadow-related code would be the
obvious guess, but the earlier code looks like it had the
same problem.

The one-line fix is to stack shadows in the right place
immediately after adding them to the stage.

BUG=chromium-os:5786
TEST=added

Review URL: http://codereview.chromium.org/3110023