wm: Load WM_CLIENT_MACHINE properties from windows.
This is an ICCCM property on a window that contains the
hostname of the machine where the client is running.  I'm
going to use this to make sure that when we try to kill a
Chrome process that's hanging, it's actually running on the
local machine instead of over the network.  (I don't know of
any plans to support networked clients, but it seems likely
that a developer could do this, and accidentally killing a
random local process would be bad.)

BUG=chromium-os:6109
TEST=added

Review URL: http://codereview.chromium.org/3218003