wm: Drop startup background earlier.
We were previously hanging on to our actor containing the
initial contents of the root window (the boot splash image)
until the first browser window was visible, which meant that
the root window actor could be briefly visible if Chrome got
killed before then due to the user shutting down during
login.

Note that the splash image is still visible when the UI
process stops even with this change because it's still
present on the framebuffer, which is briefly visible after X
exits.  I have a separate change to add framebuffer-clearing
support to ply-image, and will probably make the window
manager clear it after it's grabbed the root window from X.

What a mess.

I've also made LoginController wait for the login
background's initial pixmap to get loaded instead of for
Chrome to set a property on the window -- with the property,
we were jumping the gun on the fade-in animation.  The
animation takes too long now that the window is actually
visible for the entirety of it, so I dropped it down from
400 ms to 200 ms.

TEST=updated existing tests and ran it
BUG=chromium-os:4279

Review URL: http://codereview.chromium.org/3267005