wm: Ping Chrome periodically to check if it's hanging.
This makes us use the _NET_WM_PING protocol to ask a Chrome
window for a reply every five seconds.  If we don't hear
back from it in four seconds, we log a message (we can
additionally send SIGABRT to the process, although I have
that hidden behind an off-by-default flag for now).

BUG=chromium-os:6109
TEST=added

Review URL: http://codereview.chromium.org/3285001