wm: Load windows' titles when they're first created.
This fixes a dumb-but-harmless (as we don't do use titles
for anything but debugging) bug where we'd only load
windows' titles in response to PropertyNotify events.  It's
possible for a client to create a window and set its title
property before we select PropertyChangeMask on it, in which
case the old code wouldn't load its title.

BUG=none
TEST=added

Review URL: http://codereview.chromium.org/3433006