wm: Walk up window tree to find transient owners.
When Flash opens a transient window (the only ones that I
know of are file pickers), it appears to assume that the
window embedding the Flash object is the toplevel browser
window.  Unfortunately, this isn't the case in the Linux
version of Chrome, where the Flash window is nested several
layers down the tree.  As such, we end up not being able to
find the transient window's owner, and we handle this pretty
poorly currently; the transient window ends up mapped and
onscreen, but its actor is invisible.

This change makes LayoutManager walk up the window tree when
it sees a transient claim to be owned by a non-toplevel
window, so that it can find the actual toplevel window.

BUG=chromium-os:5846
TEST=added, and also tested that the filepicker at jaycut.com now works

Review URL: http://codereview.chromium.org/3450008