wm: Grab pointer/keyboard and hide cursor during shutdown.
This makes us grab the pointer and keyboard (which may fail,
if the screen is locked) and creates a transparent cursor
and assigns it to the root window.  It also cleans up some
of the cursor-related code a bit and renames some grabbing
and un-grabbing methods.

BUG=chromium-os:3382
TEST=added; also built and ran it

Review URL: http://codereview.chromium.org/3592013

Change-Id: I0119d5fb291a033b0cf0cb33b5e3ee9b19181df2