wm: Disable overview mode; add tab-cycling.
This makes the overview mode key (F5) cycle to the next tab
(cycling windows as needed), and Shift-Overview cycle to the
previous tab.  This is just a quick hack since there are a
few showstoppers in overview mode at the moment.  It should
be disabled or reverted once those are fixed.

BUG=chromium-os:7263,chromium-os:7333
TEST=added

Review URL: http://codereview.chromium.org/3702003