wm: exit(1) instead of raising SIGABRT when X dies.
This is an attempt to avoid getting reports that the window
manager has died when the crash was really caused by X going
away.  We get crash reports from X, so I'm hoping that this
won't make us miss learning about any crashes.

I'm also making us exit(1) when we fail to connect to the X
server in the first place.  I'm not sure if this is as good
of an idea, but it seems to be necessary, since the session
manager restarts the WM after X dies even if the WM returned
0.

Change-Id: I5b4700be3269b2405217b1ca6ee6bcc1c5fef862

BUG=none
TEST=built it, kill -9ed X, and checked ~/user/crash

Review URL: http://codereview.chromium.org/3744002