wm: Defer locking until screen locker window is visible.
This gets rid of some of jank in our animations.  Instead
of considering the screen to be locked (and starting the
locking animation) as soon as a screen locker window is
mapped, we wait until we've fetched the window's initial
pixmap so we can actually draw it onscreen.

I'm also adding a new stacking layer for the snapshots that
we animate when locking or shutting down (otherwise, the
screen locker window can get stacked on top of them in some
cases) and fixing a typo that I introduced in mock_chrome
(which doesn't get built by default).

BUG=chromium-os:7564
TEST=added; also built and ran it

Review URL: http://codereview.chromium.org/4044001