wm: Claim to support _NET_WM_MOVERESIZE.
This should make GDK avoid some fallback code that looks
like it contains a bug.

(This protocol lets clients send messages to the window
manager telling it to initiate a window move or resize.  It
looks like there may be a bug in the GDK code that's used if
the WM says it doesn't support this protocol, leading to
failed assertions.  We shouldn't be doing user-initiated
moves or resizes in Chrome OS, so it's unclear to me why
this is even getting triggered -- it may be a bug in Views.)

Change-Id: I6ef53c766fc1ca038b212421102ffc7cf04caced

BUG=chromium-os:8610
TEST=none (well, built and ran it and verified that Chrome now sends _NET_WM_MOVERESIZE messages when i try to drag a popped-out inspector window, but i don't know how to repro the crash)

Review URL: http://codereview.chromium.org/4552005