wm: Calibrate display if color profile is present.
This makes the window manager run xcalib to load an ICC
color profile while we're fading from the boot splash image
to the login background.  As mentioned in a TODO, this is
pretty hacky but the alternative seems worse: If we instead
load the profile as soon as X is started, there's a
noticeable downward step in brightness.  This way, the step
gets at least partially lost in the noise while the
background is changing. :-/

Change-Id: Ia8038f4bbe3f195b2c26aefb41390b0b8dfec69e

BUG=chromium-os:8744
TEST=built and ran it

Review URL: http://codereview.chromium.org/4656001