wm: Don't show animation for non-power-button screen lock.
This makes us avoid showing the fast-close animation when we
transition directly from the unlocked to locked state (i.e.
without the power button being held).  This happens when the
screen is locked just before suspending, so we want to make
sure that we're not showing the previous contents of the
screen when the system wakes up.

BUG=chromium-os:8867
TEST=added

Review URL: http://codereview.chromium.org/4636003

Change-Id: I38d0c0bda8e534df2e932ab344d8547bcb271502