wm: Focus first login entry's controls window immediately.
This makes us assign the input focus to the first login
entry's controls (password) window as soon as we map it
instead of waiting until all of the login windows have been
painted.

This removes an up-to-several-seconds period where the
window doesn't have the focus, a focus ring isn't drawn
around it, and the user's keystrokes may be dropped.

Note that Chrome appears to still take several seconds to
process the first keystrokes and update the text field as
the user is typing their password, though.  I will continue
investigating the cause of this.

BUG=chromium-os:10456
TEST=added; also checked that focus is assigned as expected in the OOBE, single-sign-up-pod, and multiple-user-pods cases

Review URL: http://codereview.chromium.org/6008003

Change-Id: I8abab4e6e1460e8421615997d1b5720fc9d87ee6