wm: Support showing info bubbles while screen is locked.
This makes the window manager honor a property that Chrome
sets on volume and brightness bubbles stating that they
should be visible while the screen is locked.

Things still look a bit janky if you hit a volume key and
immediately lock the screen, since the volume bubble is
present both in the snapshot of the screen that we're
scaling down and on top of the screen locker window that's
behind the snapshot.  This can be fixed at some later point
by either scaling the actual windows instead of taking a
snapshot (which would require some
transform-a-subset-of-actors-regardless-of-stacking feature
in the compositor) or by temporarily drawing a frame where
the bubble is hidden before taking the snapshot, but it
doesn't seem like a big deal for now.

BUG=chromium-os:8232
TEST=built and ran it

Review URL: http://codereview.chromium.org/5971004

Change-Id: I1bb93c6caac7e8683e5b72e2aaa9f637bccd1e01