wm: Don't switch to new windows when modal dialog is open.
Chrome blocks the Ctrl-N accelerator while a modal dialog is
up but it still was possible to hit Ctrl-Alt-T to get the
window manager to open a terminal, which we would switch to
but display in a dimmed state.

This change makes us not switch to new toplevel windows
while a modal dialog is up, which has the effect of making
Chrome's dialogs suddenly become OS-modal (whwich seems
reasonable since we already disable window-switching
accelerators like Alt-Tab while a modal dialog is up).

Change-Id: I2ccde44bb6d6bef6b2670ac2dfdb902db0121f3b

BUG=chromium-os:10546
TEST=updated existing tests to check new behavior and ran them; also installed on a device and checked that we no longer switch to terminals opened with Ctrl-Alt-T while a modal dialog is open

Review URL: http://codereview.chromium.org/6049004