wm: Repeatedly try to grab inputs when session ends.
The window manager tries to grab the pointer and the
keyboard when the session is ending to hide the cursor and
capture any stray input.  This change makes us repeatedly
retry grabbing the devices if we fail (typically due to
Chrome already having them grabbed).

Change-Id: I52d5c76635c2ae53f8c261dbcf43dc97321694f6

BUG=chromium-os:10628
TEST=added one

Review URL: http://codereview.chromium.org/6050013