wm: Hide cursor until user moves the pointer.
This makes us grab the pointer at startup and release it once
the user moves it or the first browser window is shown
onscreen.  Grabbing the pointer lets us assign it a
transparent cursor that won't be visible when it's plopped on
top of the boot splash screen's logo.

BUG=chromium-os:8406
TEST=added.  booted a machine with this change a bunch of times, and also checked that the pointer doesn't get grabbed if we're restarted post-login

Review URL: http://codereview.chromium.org/6093005

Change-Id: I8be01ee861818930e030be777e2fa5f578428481