wm: Lay groundwork for multiple-frame animations.
This gets rid of the ugly templatized code that I added to
RealCompositor earlier and replaces it with a float-specific
Animation class that supports multiple keyframes (as opposed
to just a starting and ending frame).

Note that I'm not exposing any way to use custom animations
in this change.  I'll do that in a later one.

BUG=chromium-os:10301
TEST=added, also tested it on a device in conjunction with the aforementioned later change to support adding custom animations

Review URL: http://codereview.chromium.org/6173003

Change-Id: Id48dbb9535c3c7303ec582ea3dcff4877c9be163