wm: Update variables in comments from 'foo' to |foo|.
All of this code was originally written with single quotes
around variable names when they appear in comments, style,
but the tendency in the Chrome codebase seems to be to use
pipes instead.  This updates all of the window manager code
to use pipes as well.

Change-Id: I3e704f00a85970450dd5df0598791a479b692db8

BUG=none
TEST=built it
TBR=gspencer

Review URL: http://codereview.chromium.org/6175004