wm: Constrain panels' transient windows onscreen.
There's no issue with modal dialogs belonging to browser
windows, since we switch to the window when its dialog is
mapped.  We don't have anything implemented to deal with
panel overflow yet (http://crosbug.com/724), so if an
offscreen panel opens a modal dialog, the user might not
even be able to see it.

This change adds a setting to TransientWindowCollection to
force transient windows to remain onscreen and enables it
for panels' collections.

Change-Id: Id005f0a00ab24abddd9a436bd79afb034f2959e4

BUG=chromium-os:9054
TEST=added; also dragged a panel partially offscreen and checked that alerts opened via the JS console appeared entirely onscreen

Review URL: http://codereview.chromium.org/6312084