wm: Handle simultaneous move/resizes of windows.
Resizing a window while keeping the right or bottom edges in
the same position is tricky in X.  The move and resize happen
atomically in the X server, but the window can't be
composited at its new size until the client has finished
painting it.

This change makes the Window class take the difference
between a window's configured size and its current pixmap's
size into account to handle this without looking janky.  In
the move-and-resize case described above, we defer moving the
actor to the new position until we've loaded the resized
pixmap.

BUG=chromium-os:11519
TEST=added; also resized a panel by dragging its left edge and checked that its right edge no longer jumps for a single frame when the drag ends

Review URL: http://codereview.chromium.org/6246049

Change-Id: I2ec215693877e6ca00dbb71c4308d639d34d485c