wm: Handle client destroying window while we're mapping it.
This fixes a race that has shown up in some tests.  A client
could send a request to map a window and then immediately
destroy it; if the window was destroyed before our forwarded
map request made it to the X server, we'd incorrectly treat
the window as mapped and end up failing an assert later due
to us not seeing an unmap notify event before the destroy
notify.

This change makes us wait for the X server to tell us
whether the map succeeded or failed.

BUG=chromium-os:12212
TEST=added.  also repro-ed the crash by logging in as a guest with alt-b under qemu, and then checked that the crash no longer occurs with this change

Review URL: http://codereview.chromium.org/6575019

Change-Id: I536072b4ae5d2761c7f2142cf050be632fd55cac