wm: Try to clean up management of X vs. composited windows.
Users of the Window class have had to manage getting both
the composited and X (a.k.a. "client") windows to the
correct position, hiding the X window offscreen to disable
mouse input, etc.  This adds a new set of methods to
the Window class to set its visibility ("show it onscreen",
"show it onscreen but don't let it receive mouse events",
and "neither show it nor let it receive events") and a
Move() method that can then be used to move both the
composited and X window appropriately.

I'm not porting any of the existing code over to the new
methods in this change, but I may update some of it in
another change.

This change also fixes a few things that resulted in
compilation errors when I tried to build outside of the
chroot on an Ubuntu system recently.

BUG=none
TEST=added; also built an image, used it for a bit, and didn't notice anything broken (this is intended to be a no-op change)

Review URL: http://codereview.chromium.org/6646030

Change-Id: Ic02fad6bc785d2122d1b4a34dcc1d191803761c7