wm: Add Ctrl-Alt-Shift-D for visualizing damage events.
This toggles flashing regions of the screen whenever they
get redrawn by Chrome.  The more a region gets redrawn, the
more opaque the box that gets drawn on top of it is.

I'm also moving the X-window-position-debugging hotkey to
Ctrl-Alt-Shift-W (formerly Ctrl-Alt-D) and the profiling
hotkey to Ctrl-Alt-Shift-P (formerly Ctrl-Alt-P).

BUG=chromium-os:12997
TEST=tried it

Review URL: http://codereview.chromium.org/6695039

Change-Id: Icb5c187222584ca7f87cd27b55c1c8f24d9f99d7