wm: Add _CHROME_FREEZE_UPDATES property.
This adds a property that, while set on a window, tells the
window manager to avoid fetching its pixmap.  We should be
able to use this to defer compositing windows when they're
first mapped until they've been painted for the first time.
The property is based on Owen Taylor's proposal at
http://mail.gnome.org/archives/wm-spec-list/2009-June/msg00002.html.

BUG=chromium-os:11514,chromium-os:8537
TEST=added; also wrote a small X program that uses this property

Review URL: http://codereview.chromium.org/6739001

Change-Id: Ieb8b426c00778e6e837438bbad5387be0a063f2e