wm: Use Xlib (instead of XCB) for Composite extension.
I'm switching us to (mostly) use Xlib instead of the XCB
version in the hopes that it'll fix an error that we're
seeing from CompositeGetOverlayWindow when the X server is
going away during initialization.

I don't forsee any performance impact as a result of this.
Getting the overlay window and fetching pixmaps for windows
were already synchronous, and the previously-async calls to
redirect and unredirect windows happen infrequently.

Change-Id: I050a58c61cbc58c82782ab78b0551f5af6b6367b

BUG=chromium-os:13678
TEST=built and ran it; checked that things behave the same as before when redirection gets turned on and off

Review URL: http://codereview.chromium.org/6737033