wm: Refactor Shadow class into ImageGrid.
I'm going to need scalable 3x3 grids of images for drawing
translucent resize guides that look nicer than the simple
solid rectangles that we use currently.  This is pretty
close to what we use for drop shadows, so I've refactored
that code so that it can be used elsewhere.

BUG=none
TEST=added a bunch of tests, also built and ran it and checked that shadows look the same as before

Review URL: http://codereview.chromium.org/6810036

Change-Id: Iad8edef6238efd2c5bfa3b43ae4f14682c2ba2cd